+++ app/engine/agent.py
@@ -1,10 +1,11 @@
 from app.core.config import get_llm
 from app.engine.prompts import get_rag_prompt
 from app.index.vector_store import load_vectorstore
-from langchain_core.runnables import Runnable
+from langchain_core.runnables import Runnable,RunnablePassthrough
+from langchain_core.output_parsers import StrOutputParser
 
-
-
+def format_docs(docs):
+    return "\n\n".join(doc.page_content for doc in docs)
 
 def build_rag_chain() -> Runnable:
 
@@ -13,7 +14,7 @@ def build_rag_chain() -> Runnable:
     prompt= get_rag_prompt()
 
     chain = (
-        {"context": retriever, "question": RunnablePassthrough()}
+        {"context": retriever | format_docs, "question": RunnablePassthrough()}
         | prompt
         | llm
         | StrOutputParser()